+++ visual_context_router/__init__.py
@@ -3,20 +3,23 @@
 from .core import (
     ImageInfo,
     Observation,
+    RouteDecision,
     TokenEstimate,
     UIElement,
     crop_roi,
     estimate_tokens,
     observe,
+    route_observation,
 )
 
 __all__ = [
     "ImageInfo",
     "Observation",
+    "RouteDecision",
     "TokenEstimate",
     "UIElement",
     "crop_roi",
     "estimate_tokens",
     "observe",
+    "route_observation",
 ]
-